test commit

+++ src/Controller/AccountAddressController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Address;
 use App\Form\AddressType;
+use App\Manager\CartManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -25,7 +26,7 @@ class AccountAddressController extends AbstractController
     }
 
     #[Route('/compte/ajouter-une-adresse', name: 'account_address_add')]
-    public function add(Request $request): Response
+    public function add(CartManager $cart, Request $request): Response
     {
         $address = new Address();
 
@@ -37,7 +38,12 @@ class AccountAddressController extends AbstractController
             $address->setUser($this->getUser());
             $this->entityManager->persist($address);
             $this->entityManager->flush();
-             return $this->redirectToRoute('account_address');
+            if ($cart->getCurrentCart()) {
+                return $this->redirectToRoute('order');
+            } else {
+
+                return $this->redirectToRoute('account_address');
+            }
         }
 
         return $this->render('account/address_add.html.twig', [
@@ -72,12 +78,11 @@ class AccountAddressController extends AbstractController
     {
         $address = $this->entityManager->getRepository(Address::class)->findOneById($id);
 
-        if ($address ||$address->getUser() == $this->getUser()) {
+        if ($address && $address->getUser() == $this->getUser()) {
             $this->entityManager->remove($address);
             $this->entityManager->flush();
         }   
-
-           
+        
             return $this->redirectToRoute('account_address');
